backend/ocr_service/requirements.txt
fastapi==0.114.2
pydantic==2.9.2
uvicorn[standard]==0.30.6
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10
gunicorn==22.0.0
python-multipart==0.0.9
